64 bits warnings fix
git-svn-id: http://ltt.polymtl.ca/svn@2738 04897980-b3bd-0310-b5e0-8ef037075253

+++ ltt/branches/poly/lttv/lttv/tracecontext.c
@@ -686,7 +686,7 @@ void lttv_process_traceset_begin(LttvTracesetContext *self,
 /* Note : a _middle must be preceded from a _seek or another middle */
 guint lttv_process_traceset_middle(LttvTracesetContext *self,
                               LttTime end,
-                              guint nb_events,
+                              gulong nb_events,
                               const LttvTracesetContextPosition *end_position)
 {
   GTree *pqueue = self->pqueue;
